--- a/spec/features/staff/users_can_create_a_third_party_project_level_activity_spec.rb
+++ b/spec/features/staff/users_can_create_a_third_party_project_level_activity_spec.rb
@@ -35,6 +35,9 @@ RSpec.feature "Users can create a third-party project" do
         third_party_project = project.child_activities.last
 
         expect(third_party_project.organisation).to eq user.organisation
+
+        activity = Activity.order("created_at ASC").last
+        expect(activity.transparency_identifier).to eql("GB-GOV-13-#{activity.roda_identifier}")
       end
 
       context "without an editable report" do
@@ -49,24 +52,6 @@ RSpec.feature "Users can create a third-party project" do
 
           expect(page).to have_no_button t("action.activity.add_child")
         end
-      end
-
-      scenario "the activity saves its identifier as read-only `transparency_identifier`" do
-        programme = create(:programme_activity, :gcrf_funded, extending_organisation: user.organisation)
-        project = create(:project_activity, :gcrf_funded, organisation: user.organisation, extending_organisation: user.organisation, parent: programme)
-        _report = create(:report, state: :active, organisation: user.organisation, fund: project.associated_fund)
-
-        visit activities_path
-
-        click_on(project.title)
-        click_on t("tabs.activity.children")
-
-        click_on(t("action.activity.add_child"))
-
-        fill_in_activity_form(level: "third_party_project", parent: project)
-
-        activity = Activity.order("created_at ASC").last
-        expect(activity.transparency_identifier).to eql("GB-GOV-13-#{activity.roda_identifier}")
       end
 
       scenario "a new third party project requires specific fields when the project is Newton-funded" do