@pezholio 2021-08-24
![](https://github.trello.services/images/mini-trello-icon.png) [Switch out the aid type features for requests](https://trello.com/c/4E9H37d7/2128-switch-out-the-aid-type-features-for-requests)